Remove DataSetDownloader
- POC no longer requires dataset information (footfall, urbanization, landtype)
- Removed rewarder check for unprocessed datasets
- Removed settings related to data set downloading
- Deleted boosting_oracles test suite
- Deleted modeled_coverage test suite
- Moved location_trust_score test to heartbeats suite
- Removed unused mock code related to DataSets
- MIGRATION: delete hex_assignments_data_set_status table,
    data_set_status type, data_set_type type.
- AssignedCoverageObject streams and support left
  - Will be removed when old poc reward path fully torn out
