+++ mobile_verifier/Cargo.toml
@@ -8,8 +8,6 @@ authors.workspace = true
 
 [dependencies]
 tls-init = { path = "../tls_init" }
-async-compression = { version = "0", features = ["tokio", "gzip"] }
-regex = "1"
 anyhow = { workspace = true }
 async-trait = { workspace = true }
 base64 = { workspace = true }
@@ -44,7 +42,6 @@ strum_macros = { workspace = true }
 thiserror = { workspace = true }
 tokio = { workspace = true }
 tokio-stream = { workspace = true }
-tokio-util = { workspace = true }
 tonic = { workspace = true }
 tracing = { workspace = true }
 tracing-subscriber = { workspace = true }
@@ -72,4 +69,3 @@ trino-rust-client = { version = "0.9" }
 
 [dev-dependencies]
 proptest = "1.5.0"
-tempfile = "3"
