+++ mobile_verifier/src/rewarder.rs
@@ -1,7 +1,5 @@
 use crate::{
-    banning,
-    boosting_oracles::db::check_for_unprocessed_data_sets,
-    coverage,
+    banning, coverage,
     data_session::{self, DataSessionSource},
     heartbeats::{self, HeartbeatReward},
     iceberg, resolve_subdao_pubkey,
@@ -227,11 +225,6 @@ where
 
             if db::no_data_transfer_sessions(&self.data_session_source, reward_period).await? {
                 tracing::info!("No Burned Data Transfer Sessions found past reward period");
-                return Ok(false);
-            }
-
-            if check_for_unprocessed_data_sets(&self.pool, reward_period.end).await? {
-                tracing::info!("Data sets still need to be processed");
                 return Ok(false);
             }
         } else {
