+++ mobile_verifier/tests/integrations/main.rs
@@ -2,12 +2,10 @@ mod common;
 
 mod banning;
 mod banning_iceberg;
-mod boosting_oracles;
 mod coverage;
 mod heartbeats;
 mod heartbeats_iceberg;
 mod last_location;
-mod modeled_coverage;
 mod reward_dc_hip_149;
 mod rewarder_dc_trino;
 mod rewarder_poc_dc;
