+++ mobile_verifier/tests/integrations/reward_dc_hip_149.rs
@@ -1,19 +1,11 @@
 //! Integration tests for the HIP-149 reward path (`reward_dc_hip_149`): data
 //! transfer consumes the entire emissions pool and Proof-of-Coverage is not
-//! rewarded. These mirror the seeding of the old `rewarder_poc_dc` tests
-//! (heartbeats / speedtests / coverage / bans / unique connections) precisely to
-//! prove the new path ignores every PoC input and only ever emits DC rewards
-//! plus a single rounding-dust `UnallocatedReward`.
-
-use std::ops::Range;
+//! rewarded. Each hotspot with burned data-transfer sessions splits the pool in
+//! proportion to its data credits, with rounding dust emitted as a single
+//! `UnallocatedReward`.
 
 use crate::common::{self, default_price_info, reward_info_24_hours};
 use chrono::{DateTime, Duration as ChronoDuration, Utc};
-use file_store_oracles::{
-    coverage::{CoverageObject as FSCoverageObject, KeyType, RadioHexSignalLevel},
-    speedtest::CellSpeedtest,
-    unique_connections::{UniqueConnectionReq, UniqueConnectionsIngestReport},
-};
 use helium_crypto::PublicKeyBinary;
 use helium_iceberg::IcebergTestHarness;
 use helium_iceberg_oracles::data_transfer::burned_session::{
@@ -21,23 +13,11 @@ use helium_iceberg_oracles::data_transfer::burned_session::{
 };
 use mobile_config::sub_dao_epoch_reward_info::EpochRewardInfo;
 use mobile_verifier::{
-    cell_type::CellType,
-    coverage::CoverageObject,
     data_session::{self, DataSessionSource},
-    heartbeats::{Heartbeat, ValidatedHeartbeat},
-    reward_shares, rewarder, speedtests, unique_connections,
+    reward_shares, rewarder,
 };
 use rust_decimal::prelude::*;
-use rust_decimal_macros::dec;
 use sqlx::{PgPool, Postgres, Transaction};
-use uuid::Uuid;
-
-pub mod proto {
-    pub use helium_proto::services::poc_mobile::{
-        CoverageObjectValidity, HeartbeatValidity, LocationSource, SeniorityUpdateReason,
-        SignalLevel,
-    };
-}
 
 const HOTSPOT_1: &str = "112NqN2WWMwtK29PMzRby62fDydBJfsCLkCAf392stdok48ovNT6";
 const HOTSPOT_2: &str = "11uJHS2YaEWJqgqC7yza9uvSmpv5FWoMQXiP8WbxBGgNUmifUJf";
@@ -104,67 +84,6 @@ async fn test_dc_rewards(pool: PgPool) -> anyhow::Result<()> {
             "unallocated should never exceed an individual gateways rewards"
         );
     }
-
-    Ok(())
-}
-
-#[sqlx::test]
-async fn test_poc_inputs_are_ignored(pool: PgPool) -> anyhow::Result<()> {
-    let (mobile_rewards_client, mobile_rewards) = common::create_file_sink();
-
-    let reward_info = reward_info_24_hours();
-
-    let pubkey: PublicKeyBinary = HOTSPOT_3.to_string().parse().unwrap(); // wifi hotspot
-
-    let harness = common::setup_iceberg().await?;
-
-    // Seed every input that *would* earn coverage rewards on the old path:
-    // heartbeats, speedtests, good oracle assignments, and a qualifying
-    // unique-connections report — alongside the data sessions. The HIP-149 path
-    // reads none of them, so a fully-qualified radio still earns no PoC.
-    let mut txn = pool.clone().begin().await?;
-    seed_heartbeats(reward_info.epoch_period.start, &mut txn).await?;
-    seed_speedtests(reward_info.epoch_period.end, &mut txn).await?;
-    let rewardable_total =
-        seed_data_sessions(reward_info.epoch_period.start, &mut txn, &harness).await?;
-    txn.commit().await?;
-    update_assignments(&pool).await?;
-
-    let mut txn = pool.begin().await?;
-    seed_unique_connections(&mut txn, &[(pubkey.clone(), 42)], &reward_info.epoch_period).await?;
-    txn.commit().await?;
-
-    let trino = trino_client::Client::from_client(harness.owned_trino().await?);
-
-    rewarder::reward_dc_hip_149(
-        &DataSessionSource::new(pool.clone(), Some(trino)),
-        mobile_rewards_client,
-        &reward_info,
-        default_price_info(),
-        None,
-    )
-    .await?;
-
-    let rewards = mobile_rewards.finish().await?;
-
-    // No coverage reward of any kind is emitted, despite fully-qualifying inputs.
-    assert!(rewards.radio_reward_v2s.is_empty());
-    assert!(rewards.radio_rewards.is_empty());
-
-    // DC allocation is unaffected: all three hotspots are rewarded.
-    let dc_rewards = &rewards.gateway_rewards;
-    assert_eq!(dc_rewards.len(), 3);
-
-    // rewardable_bytes (not upload + download) is what flows through to rewards.
-    let rewardable_sum: u64 = dc_rewards.iter().map(|r| r.rewardable_bytes).sum();
-    assert_eq!(rewardable_total, rewardable_sum);
-
-    // The whole data-transfer pool is accounted for.
-    let expected_sum = expected_data_transfer_pool(&reward_info);
-    assert_eq!(
-        rewards.dc_transfer_sum() + rewards.unallocated_sum(),
-        expected_sum
-    );
 
     Ok(())
 }
@@ -481,158 +400,6 @@ async fn test_backstop_grows_data_transfer_pool(pool: PgPool) -> anyhow::Result<
     Ok(())
 }
 
-async fn seed_heartbeats(
-    ts: DateTime<Utc>,
-    txn: &mut Transaction<'_, Postgres>,
-) -> anyhow::Result<()> {
-    for n in 0..24 {
-        let hotspot_key1: PublicKeyBinary = HOTSPOT_1.to_string().parse().unwrap();
-        let cov_obj_1 = create_coverage_object(
-            ts + ChronoDuration::hours(n),
-            hotspot_key1.clone(),
-            0x8a1fb466d2dffff_u64,
-            true,
-        );
-        let wifi_heartbeat_1 = ValidatedHeartbeat {
-            heartbeat: Heartbeat {
-                hotspot_key: hotspot_key1,
-                operation_mode: true,
-                lat: 0.0,
-                lon: 0.0,
-                coverage_object: Some(cov_obj_1.coverage_object.uuid),
-                location_validation_timestamp: None,
-                timestamp: ts + ChronoDuration::hours(n),
-                heartbeat_timestamp: ts + ChronoDuration::hours(n),
-                location_source: proto::LocationSource::Gps,
-            },
-            cell_type: CellType::NovaGenericWifiIndoor,
-            distance_to_asserted: Some(0),
-            asserted_location: None,
-            device_type: None,
-            coverage_meta: None,
-            location_trust_score_multiplier: dec!(1.0),
-            validity: proto::HeartbeatValidity::Valid,
-        };
-
-        let hotspot_key2: PublicKeyBinary = HOTSPOT_2.to_string().parse().unwrap();
-        let cov_obj_2 = create_coverage_object(
-            ts + ChronoDuration::hours(n),
-            hotspot_key2.clone(),
-            0x8a1fb49642dffff_u64,
-            true,
-        );
-        let wifi_heartbeat_2 = ValidatedHeartbeat {
-            heartbeat: Heartbeat {
-                hotspot_key: hotspot_key2,
-                operation_mode: true,
-                lat: 0.0,
-                lon: 0.0,
-                coverage_object: Some(cov_obj_2.coverage_object.uuid),
-                location_validation_timestamp: None,
-                timestamp: ts + ChronoDuration::hours(n),
-                heartbeat_timestamp: ts + ChronoDuration::hours(n),
-                location_source: proto::LocationSource::Gps,
-            },
-            cell_type: CellType::NovaGenericWifiIndoor,
-            distance_to_asserted: Some(0),
-            asserted_location: None,
-            device_type: None,
-            coverage_meta: None,
-            location_trust_score_multiplier: dec!(1.0),
-            validity: proto::HeartbeatValidity::Valid,
-        };
-
-        let hotspot_key3: PublicKeyBinary = HOTSPOT_3.to_string().parse().unwrap();
-        let cov_obj_3 = create_coverage_object(
-            ts + ChronoDuration::hours(n),
-            hotspot_key3.clone(),
-            0x8c2681a306607ff_u64,
-            true,
-        );
-        let wifi_heartbeat_3 = ValidatedHeartbeat {
-            heartbeat: Heartbeat {
-                hotspot_key: hotspot_key3,
-                operation_mode: true,
-                lat: 0.0,
-                lon: 0.0,
-                coverage_object: Some(cov_obj_3.coverage_object.uuid),
-                location_validation_timestamp: Some(ts - ChronoDuration::hours(24)),
-                timestamp: ts + ChronoDuration::hours(n),
-                heartbeat_timestamp: ts + ChronoDuration::hours(n),
-                location_source: proto::LocationSource::Skyhook,
-            },
-            cell_type: CellType::NovaGenericWifiIndoor,
-            distance_to_asserted: Some(0),
-            asserted_location: None,
-            device_type: None,
-            coverage_meta: None,
-            location_trust_score_multiplier: dec!(1.0),
-            validity: proto::HeartbeatValidity::Valid,
-        };
-
-        save_seniority_object(ts + ChronoDuration::hours(n), &wifi_heartbeat_3, txn).await?;
-        save_seniority_object(ts + ChronoDuration::hours(n), &wifi_heartbeat_1, txn).await?;
-        save_seniority_object(ts + ChronoDuration::hours(n), &wifi_heartbeat_2, txn).await?;
-
-        wifi_heartbeat_1.save(txn).await?;
-        wifi_heartbeat_2.save(txn).await?;
-        wifi_heartbeat_3.save(txn).await?;
-
-        cov_obj_1.save(txn).await?;
-        cov_obj_2.save(txn).await?;
-        cov_obj_3.save(txn).await?;
-    }
-    Ok(())
-}
-
-async fn update_assignments(pool: &PgPool) -> anyhow::Result<()> {
-    let _ = common::set_unassigned_oracle_boosting_assignments(
-        pool,
-        &common::mock_hex_boost_data_default(),
-    )
-    .await?;
-    Ok(())
-}
-
-async fn seed_speedtests(
-    ts: DateTime<Utc>,
-    txn: &mut Transaction<'_, Postgres>,
-) -> anyhow::Result<()> {
-    for n in 0..24 {
-        let hotspot1_speedtest = CellSpeedtest {
-            pubkey: HOTSPOT_1.parse().unwrap(),
-            serial: "serial1".to_string(),
-            timestamp: ts - ChronoDuration::hours(n * 4),
-            upload_speed: 100_000_000,
-            download_speed: 100_000_000,
-            latency: 50,
-        };
-
-        let hotspot2_speedtest = CellSpeedtest {
-            pubkey: HOTSPOT_2.parse().unwrap(),
-            serial: "serial2".to_string(),
-            timestamp: ts - ChronoDuration::hours(n * 4),
-            upload_speed: 100_000_000,
-            download_speed: 100_000_000,
-            latency: 50,
-        };
-
-        let hotspot3_speedtest = CellSpeedtest {
-            pubkey: HOTSPOT_3.parse().unwrap(),
-            serial: "serial3".to_string(),
-            timestamp: ts - ChronoDuration::hours(n * 4),
-            upload_speed: 100_000_000,
-            download_speed: 100_000_000,
-            latency: 50,
-        };
-
-        speedtests::save_speedtest(&hotspot1_speedtest, txn).await?;
-        speedtests::save_speedtest(&hotspot2_speedtest, txn).await?;
-        speedtests::save_speedtest(&hotspot3_speedtest, txn).await?;
-    }
-    Ok(())
-}
-
 /// A logical data-transfer session, materializable into either backend so the
 /// same input can drive the Postgres and Trino reward paths.
 struct DataSession {
@@ -770,81 +537,4 @@ fn data_sessions_with_dc(ts: DateTime<Utc>, dcs: &[(&str, u64)]) -> Vec<DataSess
             timestamp,
         })
         .collect()
-}
-
-async fn seed_unique_connections(
-    txn: &mut Transaction<'_, Postgres>,
-    things: &[(PublicKeyBinary, u64)],
-    epoch: &Range<DateTime<Utc>>,
-) -> anyhow::Result<()> {
-    let mut reports = vec![];
-    for (pubkey, unique_connections) in things {
-        reports.push(UniqueConnectionsIngestReport {
-            received_timestamp: epoch.start + chrono::Duration::hours(1),
-            report: UniqueConnectionReq {
-                pubkey: pubkey.clone(),
-                start_timestamp: Utc::now(),
-                end_timestamp: Utc::now(),
-                unique_connections: *unique_connections,
-                timestamp: Utc::now(),
-                carrier_key: pubkey.clone(),
-                signature: vec![],
-            },
-        });
-    }
-    unique_connections::db::save(txn, &reports).await?;
-    Ok(())
-}
-
-fn create_coverage_object(
-    ts: DateTime<Utc>,
-    pub_key: PublicKeyBinary,
-    hex: u64,
-    indoor: bool,
-) -> CoverageObject {
-    let location = h3o::CellIndex::try_from(hex).unwrap();
-    let key_type = KeyType::HotspotKey(pub_key.clone());
-    let report = FSCoverageObject {
-        pub_key,
-        uuid: Uuid::new_v4(),
-        key_type,
-        coverage_claim_time: ts,
-        coverage: vec![RadioHexSignalLevel {
-            location,
-            signal_level: proto::SignalLevel::High,
-            signal_power: 1000,
-        }],
-        indoor,
-        trust_score: 1000,
-        signature: Vec::new(),
-    };
-    CoverageObject {
-        coverage_object: report,
-        validity: proto::CoverageObjectValidity::Valid,
-    }
-}
-
-//TODO: use existing save methods instead of manual sql
-async fn save_seniority_object(
-    ts: DateTime<Utc>,
-    hb: &ValidatedHeartbeat,
-    exec: &mut Transaction<'_, Postgres>,
-) -> anyhow::Result<()> {
-    sqlx::query(
-        r#"
-        INSERT INTO seniority
-          (radio_key, last_heartbeat, uuid, seniority_ts, inserted_at, update_reason, radio_type)
-        VALUES
-          ($1, $2, $3, $4, $5, $6, 'wifi'::radio_type)
-        "#,
-    )
-    .bind(hb.heartbeat.key())
-    .bind(hb.heartbeat.timestamp)
-    .bind(hb.heartbeat.coverage_object)
-    .bind(ts)
-    .bind(ts)
-    .bind(proto::SeniorityUpdateReason::NewCoverageClaimTime as i32)
-    .execute(&mut **exec)
-    .await?;
-    Ok(())
 }
